Storage vault (#20)
